Gráficos utilizando datos del Json , falta el gráfico de Plataformas
