+++ seccion_estadisticas_v2/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-//import 'package:seccion_estadisticas_v2/pruebas_pie_chart.dart';
-import 'package:seccion_estadisticas_v2/stadistics_screen.dart';
+import 'package:seccion_estadisticas_v2/statistics_screen.dart';
 
 void main() {
   runApp(const MainApp());
@@ -11,6 +10,6 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(home: StadisticsScreen());
+    return const MaterialApp(home: StatisticsScreen());
   }
 }
